Add files via upload
This Repsitory contains the files for the GWU Bootcamp Module 3 Python Challenge assignments:  PyBank & PyPoll.
For each part (PyBank and PyPoll), there are 4 associated files:
1)  a csv file with the incoming data (budget_data.csv for PyBank, election_data.csv for PyPoll).
2)  a text file with the analytic output (budget_analysis.txt  for PyBank, election_analysis.txt for PyPoll).
      Note, both of these text files are programmatically output within th Python code.
3)  the python code, both named 'main.py', but uploaded from separate project directories.
4)  a text file with the python script (just for extra safe-keeping ;) ).

In completing this assignment, I went to class office hours, engaged with fellow students, and attended a tutoring session.  It was a valuable experience.  And, Python is no longer a 'great unknown' :).